Disallow 'balance' disambiguation in arithmetic methods
We could not think of a use case for this disambiguation mode in the
arithmetic methods. In with() and from() it's still useful, but in
addition and subtraction it just seems to be confusing.

We consider arithmetic disambiguation, like disambiguation when
converting between Absolute and DateTime, to be a separate operation.

Closes: #344.