Add limits to calendar-based types
- DateTime's limits are exactly enough to represent any Absolute in any
  valid TimeZone (offset -24 to +24 hours, exclusive at both
  boundaries.)
- Date's limits are exactly enough that a call of getDate() on any valid
  DateTime succeeds.
- YearMonth's limits are exactly enough that a call of getYearMonth() on
  any valid Date succeeds.

This also means that the inverse conversions of those mentioned above
can fail. If they are called in 'constrain' disambiguation mode, then
they'll be clamped to the nearest in-range value for the resulting type.
The 'balance' mode can't do anything in this case, so it'll behave like
'reject'.

Also ensures that plus() and minus() can't have results that are out of
the range.

See: #24.