Fix Duration.toString for sub-second units > 999
Previously, the code assumed that the millisecond, microsecond, and
nanosecond fields of Duration all had 3 digits or less. This allows for
the possibility of values ≥ 1000. In order to allow this, roundtripping
via toString() is no longer preserved, we have to balance the sub-second
units with a new abstract operation BalanceSubSecond.

Closes: #396