Disallow negative numbers as Duration fields
We now disallow negative numbers as fields when creating a Duration,
regardless of disambiguation mode. (This is because negative durations
are never valid.)

For the use case where it previously made sense to use negative numbers,
balancing { hours: 1, minutes: -30 }, the solution will be to use
Duration's arithmetic methods which will be added in the next commit.

See: #408