Collect code coverage from test262 runs
This is fraught with problems, because test262-harness concatenates
sources together, thereby ignoring the source map of the transpiled
script.js.

It's not possible to run test262 while importing index.js as a CJS
module with require(), because test262-harness executes everything in a
different realm, so e.g. assert.throws(RangeError, ...) will not work
since it will compare RangeErrors from two different realms.

It's also not possible to run test262 without transpiling, because
lib/initialise.js uses dynamic import, which test262-harness would need
to support in its custom context.

The solution chosen here is to pre-instrument the transpiled script.js
with nyc, and use a transformer function to have test262-harness append
some code to it that will output the coverage dumps to the right place.

The coverage dumps generated by nyc are a different format than the v8
native ones, so we have to generate two separate lcov files for the
regular and test262 test runs. We also submit them to codecov.io under
two separate "flags" so that they are reported separately as well as
combined.

(This also means the HTML reports generated by 'npm run coverage' only
reflect the coverage of the regular tests, not the test262 ones. But the
reports on codecov.io should reflect both.)