Allow subclassing Temporal.TimeZone
In order to allow subclasses to provide their own time zone identifier,
we need to remove the GetCanonicalTimeZoneIdentifier check from the
Temporal.TimeZone constructor when new.target is not Temporal.TimeZone.

Adds tests to verify that subclassing Temporal.TimeZone works and that
a simple object implementing the interface behaves correctly.