Define Node.js custom util.inspect formatters in debug mode
If __debug__ is defined to true, Temporal objects will gain a custom
util.inspect formatter. Like the Chrome devtools formatter, it will
print the class name and the toString() output between angle brackets,
and for Temporal.Duration it will print the nonzero fields.

This should not increase the bundle size because Rollup will eliminate
the dead code if __debug__ is defined to false at build time.

See: #643