Enable debug mode when running via 'npm run playground'
This should be OK because none of the rollup configs will process
lib/initialise.js. So when running via 'npm run playground' and
therefore 'node -r lib/initialise.js' we can detect this and define a
global __debug__ variable, which will be detected by the checks that
normally get replaced out.

See: #643