Change 'prepare' back to 'prepublishOnly' in polyfill build script
In 2fe9177d6b44d0288bb4a53db34943f500f0184d I changed this due to a
warning from npm, but I had misunderstood how it works and also failed
to notice that it actually runs rollup twice whenever you build. Change
it back to what it was, for the time being.