Disallow largestUnit higher than hours in Time.difference
In order to be consistent with smallestUnit everywhere, we remove the
ability to pass a largestUnit higher than 'hours' to Time.difference.
This was marginally useful but now that we have smallestUnit it isn't
worth the inconsistency and confusion that it will cause if smallestUnit
is allowed to be higher than 'hours'.

See: #827