Remove 'reject' from Duration overflow options
As in the other types, a value outside of the representative range now
always throws a RangeError, regardless of the overflow mode. The overflow
mode now only controls whether to balance or not. The available values are
'balance', for maximum balancing, and 'constrain', for balancing only
enough to constrain the values not to have mixed signs.

Closes: #910