Add 'auto' value to all largestUnit options
Since we are going to need a default for largestUnit in Duration.round()
that varies based on the other options given, we actually need to accept
an 'auto' value. We actually should have done this already, since in some
cases the default largestUnit in difference() methods depends on the value
of smallestUnit.

See: #856