Define valid time zone IDs
A valid calendar ID according to the tzdata documentation is a number of
components consisting of 1 to 14 ASCII alpha characters, dashes, dots, or
underscores each, joined by slashes. `.`, `..`, or any string starting
with a dash are not valid components. Add this to the grammar and enforce
it in custom time zones.

Etc/GMT time zones are an exception, as they may contain + or digits.
These are valid for parsing, but not for custom time zones.

See: #665