Add eras to Gregorian calendar
As part of this issue we decided that another difference between the ISO
and Gregorian calendars is that the latter has BC/AD eras. It's
indeterminate how this will exactly be standardized, as that's up to
ECMA-402, but we add it to the experimental Gregorian calendar in the
polyfill.

See: #901