Rename Calendar.dateDifference() to Calendar.dateUntil()
Calendars don't need to have two separate dateUntil() and dateSince()
methods, they can just use one of the two. Let's choose dateUntil() to be
the one to implement, because 'earlier.until(later)' seems to be the more
intuitive of the two, and also Calendar.dateDifference() was already
implemented that way.

See: #998