Implement options in ZonedDateTime.from()
This implements the overflow, offset, and disambiguation options in
ZonedDateTime.from(). Overflow is used with property bags when dealing
with out-of-range values. Offset is used to pick how the offset is
interpreted, if included. Disambiguation is used to pick the appropriate
Temporal.Instant if there are more than one and the offset is not included
or it is ignored via the offset option.

Also adds more comprehensive tests for ZonedDateTime.from().

Co-authored-by: Justin Grant <justin@justingrant.net>