Fix rounding in ZonedDateTime.toString()
The rounding in ZonedDateTime.toString() only affects time units. Unlike
the rounding in ZonedDateTime.round(), it should be exact time rounding,
so that the actual difference between the raw and rounded exact times is
not affected by time zone offset shifts.

This is the same outcome as in round() (and we add a test for both round()
and toString() to prove it) but in toString() it can be implemented with
fewer observable calls since we don't need to be able to round to units
that are affected by DST shifts.

See: #569