Introduce abstract operation TotalDurationNanoseconds
This operation gives the total number of nanoseconds in an exact
duration as a bigint, not losing precision. This will be reused in
Duration.compare().

This operation takes an "offset shift" parameter, used if there are days
that need to be balanced down to hours. If it's 0 (as is the case
everywhere in this commit) then days are always 24 hours. We will start
using nonzero values for this in Duration.compare().

See: #856