Use const aliases for built-ins' static methods
Per @ms2ger suggestion, converted all use of the Math built-in object to
use alias methods instead.

I'm assuming that this guidance extends to other built-ins too, so while
I was at it I added aliases for static methods on Number and Object too.

Legacy Date is used (constructor, static now(), and prototype methods)
but I wasn't sure if this is intentional (for SES) or not.
So I didn't make any changes to it.

There were built-in prototype methods used too, but I wasn't sure if
those needed to be replaced too so didn't change any. Here's the
prototype methods I found that were used:
* Regex.prototype.exec
* String.prototype.slice
* String.prototype.match
* Array.prototype.map
* Array.prototype.filter
* Array.prototype.includes
* Array.prototype.join
* Set.prorotype.has
* Set.prototype.delete
* Map.prototype.has
* Map.prototype.get
* Map.prototype.keys
* Map.prototype.values
* Array.prototype.some
* Array.prototype.push
* Array.prototype.unshift
* String.prototype.split

Finally, there's lots of use of prototype methods on `big-integer`, but
I assumed because as an import there's no need to change those calls.
LMK if this assumption was wrong.