Avoid unnecessary observable TimeZone.toString() calls in Intl
We already call toString() once on the time zone in extractOverrides(),
and we don't need the actual time zone object subsequently, so return the
result of toString() from extractOverrides() instead of the time zone
object.

As far as I can tell this is already done correctly in the spec text.