Sync calendar annotation format with the IETF draft
IETF is moving forward with [n-kk=vvvvv] (n=namespace, k=key, v=value)
rather than [n-kk-vvvvv]. By definition, Temporal uses the format that is
being standardized within IETF.

Update the note in the spec text as well to remove the text saying that
the format is tentative. This is the format that IETF is moving forward
with.

Closes: #1409
Closes: #1412