Remove support for using this-value as constructor in static methods
Any static method of a Temporal type that returns a new instance of the
type, should not use the this-value as the constructor. This makes most of
the CreateTemporalFooFromStatic abstract operations obsolete, so they are
removed, except the ones that are used in calendars.

As in the @@species commit, reorganize the test262 tests that verify this
change, into the temporalHelpers.js file.

See: #1439