Create options objects with null prototype
In NormalizeOptionsObject, which comes from ECMA-402 originally, when an
options parameter is undefined, an object with null prototype is created.
Make everywhere else that an options parameter is synthesized, consistent
with that, so create objects with null prototype rather than {}.

Ensure that separate objects are created and not re-used once they have
been passed in to user code.

(We continue to return plain objects whose prototype is Object.prototype
from methods such as getISOFields())

See: #1424