Rename RejectInstantRange → ValidateEpochNanoseconds
We are going to rename all the ValidateFoo operations in the spec text to
IsValidFoo, and remove the RejectFoo operations. This doesn't apply to the
polyfill because the form where throwing the exception is part of the
operation is more convenient in code.

But, still rename RejectInstantRange to ValidateEpochNanoseconds because
we are making the same Instant → EpochNanoseconds change to
IsValidEpochNanoseconds, because that is what is actually passed to the
function.