Optimize toLocaleString + Intl.DateTimeFormat
This commit speeds up toLocaleString and other operations that depend
on the polyfilled Intl.DateTimeFormat. The fix was to prevent
unnecessary creation of built-in Intl.DateTimeFormat objects, because
the constructor of that built-in class is slooooooow.

For more details about the underlying issue see:
https://bugs.chromium.org/p/v8/issues/detail?id=6528

In local testing, speedup is about 2.5x for ZDT toLocaleString calls.