Editorial: Remove dead code in tautological comparison
In every case where _mid_.[[Month]] = _end_.[[Month]], it also holds that
_mid_.[[Year]] = _end_.[[Year]]. In other words, due to the preceding
algorithm steps, _mid_ is always in the previous month, the same month, or
the following month compared to _end_. There is never a case where _mid_
and _end_ are 12 months apart.

Therefore, the second part of this comparison was unnecessary and would
never be compared due to short-circuiting. Replace it with an assertion.

Closes: #1649