Normative: Ensure an Object is returned from calendar.mergeFields()
In each case where CalendarMergeFields is called, it is later asserted in
PrepareTemporalFields that Type(_fields_) is Object. Previously the spec
text failed this assertion.

The tests check that a TypeError is thrown in this case, as well as that
the TypeError is thrown before the value is passed to any dateFromFields()
or similar method.