Add CI tests for unused code
This commit tests in CI for unused functions and exports using ts-prune.

There are two exported functions in ecmascript.ts that are only used by
tests. Because our tests haven't been ported over to TS yet, that usage
is invisible to ts-prune, so I marked both functions with a comment to
disable ts-prune errors until we migrate the test code to TS and we can
remove the `// ts-prune-ignore-next` comments