Normative: Temporal.Duration constructor throws if given a non-integer
The Temporal.Duration constructor, unlike Temporal.Duration.from(), would
previously drop the fractional part of numbers given to it. from() throws
if given a non-integer number, because Temporal.Duration.from('PT1.1H')
results in 1 hour, 6 minutes, but Temporal.Duration.from({ hours: 1.1 })
cannot be exact, yet should not silently drop the .1, as discussed in
issue #938.

Temporal.Duration(0, 0, 0, 0, 1.1) should now throw the same exception as
Temporal.Duration.from({ hours: 1.1 }) in order to be consistent and avoid
the surprise to users of silently dropping the .1.

Closes: #1832