Bring polyfill in line with ParseTemporalTimeZone operation in specification
This fixes some polyfill code that no longer applies. Ever since the
change to allow UTC offsets to be rounded to minutes in ISO strings, the
`identifier !== offset` line has been incorrect, but there are no tests
for this case in our test suite.

Rename the abstract operation TemporalTimeZoneFromString to
ParseTemporalTimeZone so that it matches the spec, and simplify its logic
to match what is written in the spec as well.

I'm in the process of adding test262 tests for this, (and fixing the
existing ones -- hence the entry in expected-failures.txt) but this PR
includes a quick test in the old test suite just so that it's covered in
the meantime, and to illustrate which case would previously have failed.