Fix bad code ordering in adjustCalendarDate
Re-order code to guarantee that `calendarDate.year` is set before
calculating the number of months in that year. The old code would
break for any calendar that a) used a constant `era` like `islamic`
and b) had a variable number of months in each year like `hebrew`.
Today no ICU calendars fit in that Venn diagram, but if any are added
in the future this code will break when initializing any calendar using
an eraYear/era pair in a property bag. Better to fix it now!