Make TS happier with year/eraYear
This was originally part of tc39/proposal-temporal#1977
but I removed it because it wasn't high-priority for JS.
But it's still needed in this repo so I'm adding it back.

Why? TS complains when you assign a maybe-undefined
value to a must-not-be-undefined variable. This small
runtime change is cleaner than adding a lotta type casts.