Change MergeLargestUnitOption to match spec
This was previously done with the spread syntax, but we are going to make
a change to it, so we'll need a function instead. The spec text has the
operation MergeLargestUnitOption, so name it that.

The ObjectCreate(null) step was previously missing, so add it in order to
be compliant with the spec.

UPSTREAM_COMMIT=1660e23ae982d84ff7958c85fb44ae981e315d59