Implement full set of rounding modes in polyfill
This brings the reference polyfill up to date with the new rounding modes.

It adds a quick set of tests that only test the internal implementation,
in order to make sure that the RoundNumberToIncrement operation behaves as
intended. Full tests should test the observable outcome in test262.

UPSTREAM_COMMIT=d450be195a81228d284777969bedd9e0475400bd