Move time zone annotation parsing into ParseISODateTime
Now that annotations are an integral part of IXDTF strings, it's better to
handle them at the same time inside of ParseISODateTime. This work is
preparatory to the normative change that we will have to do to accept the
annotation format defined in IXDTF.

(IXDTF is the working abbreviation of the IETF draft; it stands for
Internet Extended Date-Time Format)

Previously we would call ParseISODateTime on a string and subsequently
ParseTemporalTimeZoneString on the same string. In addition to the benefit
mentioned above, the previous situation could also lead to confusion since
it parsed the string twice, and ParseTemporalTimeZoneString also accepts
a bare time zone identifier (even though that branch of the grammar is
not used in this case, because a bare time zone identifier can never be a
valid ISO string.)

Now, the Record returned by ParseISODateTime gains a [[TimeZone]] field
which contains another Record of the same type that is returned by
ParseTemporalTimeZoneString.

UPSTREAM_COMMIT=c410e25e4761bd9c86880219558d25074d28839f