Normative: Reject relativeTo string such as "2022-08-18T17:01Z"
`relativeTo` strings must be either valid as PlainDate strings or as
ZonedDateTime strings. A string with a Z but no bracketed time zone
annotation is neither.

We probably missed this when forbidding Z designators for Plain types, so
it was an oversight.

UPSTREAM_COMMIT=2dc20bf69065c17631347f8b7e48cae651ae8a38