Simplify ParseTemporalTimeZoneString
This was a bit of a mess. Rewrite the operation so that it tries to parse
the string as a bare time zone identifier _first_, and if that fails,
as an ISO string, throwing if the ISO string doesn't contain any time zone
(whether Z designator, UTC offset, or IANA name).

This was what the operation previously did, just much more difficult to
follow.

The reference polyfill mostly already worked like this, but move the
IANA name canonicalization step out of ParseTemporalTimeZoneString and
into its only caller, to correspond better with the spec text. This also
requires bringing the time zone ID regex better in line with the grammar's
|TimeZoneIdentifier| production by adding the IANA legacy names.

UPSTREAM_COMMIT=eec8efab983131401c2092dd26a1f9e8d049b666