Normative: Require that NanosecondsToDays doesn't flip sign
When converting a number of nanoseconds to a number of days and a
nanoseconds remainder, neither the resulting number of days nor the
nanoseconds remainder should have the opposite sign as the original number
of nanoseconds.

This could happen due to shenanigans in a custom time zone's
getOffsetNanosecondsFor or getPossibleInstantsFor methods, or a custom
calendar's dateAdd method.

See: #2357

UPSTREAM_COMMIT=e13c52d0c0e89849ccf326ac7ea6c7190fc1235d