Normative: Require NanosecondsToDays remainder less than day length
When converting a number of nanoseconds to a number of days and a
nanoseconds remainder, the remainder shouldn't be longer than the length
of the last day.

This could happen due to shenanigans in a custom time zone's
getOffsetNanosecondsFor or getPossibleInstantsFor methods, or a custom
calendar's dateAdd method.

See: #2357

UPSTREAM_COMMIT=ac69b63a5904620d0271238f028b4cee068bfcad