Polyfill: Refactor GetNamedTimeZoneDateTimeParts
For consistency with other operations in the spec, rename this function to
GetNamedTimeZone... and swap its arguments.
(This has no effect on the proposal, as it's only used in parts of the
polyfill that implement the TZDB, which is not specified. This is only for
consistency in the code.)

UPSTREAM_COMMIT=1b8940dd2f689c1634d665b4899cb2fc9d1ef7df