Avoid precision loss in AddDuration
According to the spec, the addition in AddDuration must not be performed
in the floating-point domain. Anba wrote some tests that verify this.

For days, d1 + d2 practically speaking is still OK because the result is
directly stuffed into a float64-representable integer anyway, so keep that
the same for simplicity. Perform the other additions in the BigInt domain
and pass BigInts to BalanceDuration.

UPSTREAM_COMMIT=9a0565a8cba336a6dabbf10cc58ccbf665bfe023