Polyfill: Avoid rounding errors in BalanceTime
According to the spec, the arithmetic in BalanceTime must not be performed
in the floating-point domain. Anba wrote some tests that verify this.

The big-integer library has a divmod() method, but its remainder has the
same sign as the dividend, not the divisor, so we add a helper function
NonNegativeBigIntDivmod to replace NonNegativeModulo which did the same
thing but for Numbers. It can be used in one additional place, in
RoundInstant.

UPSTREAM_COMMIT=07e68f01e16f23a46e2f28e47113ba4486c8d27e