Editorial: Remove redundant ToString from ToTemporalZonedDateTime
Due to PrepareTemporalFields a few steps above this, the `offset` property
of _fields_ is already guaranteed to be an own data property with a
primitive value, which is either a string or undefined. So, there is no
need to convert it to a string here. We can remove it because this
conversion was not observable.

UPSTREAM_COMMIT=24ebcbd80b5f1aea26b5b030d41127476fbe3dce