Avoid faulty assertion in RoundDuration about integer days
At this point in the algorithm, we need to find the number of whole years
contained within _days_. Contrary to the assertion, and to the requirement
of CreateTemporalDuration, _days_ is not an integer. The algorithm is
faulty and unimplementable as written.

However, the answer (_yearsPassed_) is still correct if we temporarily
discard the fractional part in the call to CreateTemporalDuration.

Rename the variables _daysLater_ and _daysDuration_ to _wholeDaysLater_
and _wholeDaysDuration_, respectively, to emphasize that they potentially
represent a different value from _days_.

The bug was not present in the reference code because it calculates
slightly differently with floats, instead of reals as in the spec text.
Still, rename the variables there as well. While we are here, correct the
reference code to pass a Temporal.Duration instance to dateAdd, as
specified.

Closes: #2246

UPSTREAM_COMMIT=02379da7e6a9aea1a42b339ac5eec159a3a3c32a