Normative: Use string calendar values internally everywhere possible
When creating Temporal objects with a builtin ISO 8601 calendar, we want
them to be created with a string in the internal slot, not a Calendar
instance. (For example, when parsing a calendar name out of a string or
using ISO 8601 as a default.) Therefore we don't need the
GetISO8601Calendar and GetBuiltinCalendar AOs anymore.

See: #1808

UPSTREAM_COMMIT=52f3159ba4d03343e1cfdbcee1720b03e28753b8