Normative: Avoid recalculating offset ns after GetPlainDateTimeFor
If we convert an exact time into a PlainDateTime, we're already
calculating the time zone's UTC offset for that exact time. In the case
where we need the UTC offset for another purpose in the same method, we
should not call the time zone method again to recalculate it. Instead, we
call the getOffsetNanosecondsFor method once, and pass the result to the
GetPlainDateTimeFor abstract operation.

This affects the following methods, removing an observable property Get
and observable Call to getOffsetNanosecondsFor:
- Temporal.Instant.prototype.toString
- Temporal.ZonedDateTime.prototype.toString
- Temporal.ZonedDateTime.prototype.round
- Temporal.ZonedDateTime.prototype.getISOFields

UPSTREAM_COMMIT=b7a30a11afe25d7868b74ffe9008de86e7f1890d