Normative: Don't observably iterate array in built-in calendar's fields()
When calling CalendarFields with a built-in calendar, previously we would
create a JS Array from the passed List, and pass it to %Temporal.Calendar.
prototype.fields%, which would iterate it observably. So, instead we call
CalendarDateFields when the calendar is a built-in calendar, which doesn't
do anything observable at all.

See: #2289

(TypeScript refactor: change parameter type and return type of fields()
to FieldKey[])

UPSTREAM_COMMIT=392d86e15a0f43848d979ec66bfdf4aefbd7f371