Normative: Avoid reconverting Zoned- to PlainDateTime in AddZonedDateTime
There are several places where a ZonedDateTime is converted to a
PlainDateTime, which is a user-visible time zone operation, and then the
same operation is performed again right afterwards in AddZonedDateTime.

Avoid this by making AddZonedDateTime take an optional precalculated
PlainDateTime. If we have it, we can pass it in, and avoid the second
conversion.

UPSTREAM_COMMIT=853a2122a45425e9bb9262e4e7d99ec332dfc5af