Normative: Do away with CalculateOffsetShift in Duration.compare
CalculateOffsetShift is only used in Duration.compare, when relativeTo is
a ZonedDateTime. Calling it twice performs two ZonedDateTime additions,
which are potentially user-observable method calls. If we are performing
these two additions anyway, we may as well just add each duration to
relativeTo and compare the timestamps. This saves a lot of other
user-observable calls.

It also allows removing the offset-shift parameter from
TotalDurationNanoseconds, because now it is always zero. That parameter
was confusing anyway.
At the same time we remove the days parameter as well, because now there
is no indication that days should only be 24-hour days; we push the
responsibility of deciding whether 24-hour days are appropriate to the
caller.

UPSTREAM_COMMIT=4b31c7d896e7a1f83786bbee9e697d6b701af2cc