Normative: Avoid calendar operations when adding days-only duration to PlainDate
In a few places, we called CalendarDateAdd with a days-only duration.
For days-only, it's not necessary to consult the calendar: we can just
add the days in the ISO calendar space to the ISO calendar values in the
internal slots.

(also fixes a rebase error with truncate(_fractionalDays_))
Closes: #2685

UPSTREAM_COMMIT=1ff91a9730ace0af31177afab5c57bc343a5ab2f