Normative: Fast-path differences between identical objects
If two Temporal objects have identical internal slots, then we can skip
calculating the difference with a calendar method; the difference is
always 0.

This optimization isn't necessary for PlainTime or Instant differences,
since no calendar methods are called for those. Implementations can return
early for PlainTime or Instant if they like, but it won't be reflected in
the spec text in order to keep things as simple as possible.

Note that the early return still comes after the processing of the options
object. Passing an illegal options object to until() or since() should
still throw, even if the difference is zero.

(We'll also fast-path CalendarDateUntil, but this fast path also cuts out
calls to calendar.dateFromFields().)

UPSTREAM_COMMIT=27c8876007c043982201f1939a2da3e8b696237a