Editorial: Allow passing pre-looked-up TimeZone methods
Just as in CalendarDateAdd and CalendarDateUntil, add optional parameters
to GetPossibleInstantsFor and GetOffsetNanosecondsFor which can supply the
method to be called, in case the caller has already looked it up.

This doesn't make any observable change, but it will be used for
optimizations in #2519.

UPSTREAM_COMMIT=c09dae3e5c727b69ab93c79a6f4d8b201cf2a2bc