Editorial: Allow passing pre-looked-up Calendar methods
Do the same for CalendarDateFromFields as in CalendarDateAdd and
CalendarDateUntil: add an optional parameter to CalendarFields which can
supply the method to be called, in case the caller has already looked it
up.

This doesn't make any observable change, but it will be used for
optimizations in #2519.

UPSTREAM_COMMIT=b9ee4e2b1fd20258a964f75e6b40da24e45161e1