Polyfill: Avoid overflows in GetNamedTimeZoneOffsetNanoseconds
When working with an epoch nanoseconds value at the start or end of the
range, and a UTC offset with the correct sign, it may happen that
GetNamedTimeZoneDateTimeParts produces date-time parts that cannot be
passed back into GetUTCEpochNanoseconds because they are out of range in
UTC.

Use the same optimization as in the previous commit: in the absence of
changes in UTC offset, 400 years has a constant length. So calculate the
epoch nanoseconds in UTC with the remainder of dividing the year by 400,
and then add back the number of nanoseconds in the 400 year blocks.

UPSTREAM_COMMIT=ad9997d4a860bc2b987bf37c96f4cc37f9d29ccc