Polyfill: Avoid precision loss in NanosecondsToDays
According to the spec, when we increment days in NanosecondsToDays due to
an overflow from the time component of the duration, the increment must
not be performed in the floating-point domain. Anba wrote some tests that
verify this, checking the addition of 1 beyond Number.MAX_SAFE_INTEGER.

In AddZonedDateTime the days parameter is stuffed into a Temporal.Duration
in order to pass to dateAdd user code, so it's OK to convert to float
there.

UPSTREAM_COMMIT=9d79e0c7ab8dcafa0a5bbb298a6c7bd0788f9c73