Fix: Don't call GetMethod on a string calendar
When GetMethod is used to lookup the dateAdd or dateUntil method of a
calendar, we need to account for the case where the calendar is a builtin
calendar represented by a string. This was an oversight in #2482.

In hindsight, the lookups of dateAdd and dateUntil in
BalanceDurationRelative should have been grouped together as they were in
UnbalanceDurationRelative. But since that will be addressed anyway by the
Calendar Method Records in the yet-to-be-merged PR #2519, I'm not planning
to change that here.

Closes: #2547

UPSTREAM_COMMIT=083b90e50b809e61fb37f7cd1724569b81315187