Normative: Limit duration years, months, and weeks to <2³² each
In order to prevent having to use bigint arithmetic, limit years, months,
and weeks to 32 bits each in durations.

There are more changes to the reference code than to the spec in this
commit because the upper limit now allows us to rewrite the reference
code's RoundDuration algorithm in a way that's more similar to how it was
already specified in the spec text.

UPSTREAM_COMMIT=1104cad8cc530e48f494ff108db96bccad8b780a