Polyfill: Avoid observable array iteration in a few more places
In order to be able to properly test the normative change of not
observably iterating an Array in CalendarFields, we have to eliminate a
few more places in the reference code where, contrary to the spec, array
iteration was observable.

(Rebase note: This contains many more adjustments than the original
commit. Probably these adjustments were skipped in an earlier rebase.)

UPSTREAM_COMMIT=ca7288c4a74205bb007186543199787ec3ea8ef5