Polyfill: Align HandleDateTimeTemporalTime with spec
The spec creates a PlainDateTime with the ISO 8601 calendar here, not with
the DateTimeFormat's calendar. It may make an observable difference in the
case where GetInstantFor has to disambiguate, although after #2519 that
difference should not be observable anymore. In any case, the spec says to
do this, so we should align the reference code with that.

UPSTREAM_COMMIT=6af74eef2f98140e69c6e8ef3ea5d3131c7afd5a