Normative: Limit day length calculations to safe integers
Carefully crafted custom time zones can cause NormalizedTimeDurationToDays
to calculate a day length that is arbitrarily long. In order for
implementations not to have to use a normalized time duration or bigint to
represent the day length in nanoseconds, limit it to less than 2⁵³ ns.
This way, it can be stored in a 64-bit float.

This allows time zones to specify day lengths of up to ~104.25 real
24-hour days, which should be more than enough.

This also ensures that the number of days returned from
NormalizedTimeDurationToDays is within the limit of 2⁵³/86400, which
allows implementations to use 64-bit integer addition in RoundDuration.

UPSTREAM_COMMIT=156c7644af5447a99f36dfc1c5559c1c217b7c82