Editorial: Limit year, month, and week length calculations to nonzero
In order to calculate the rounded value and totals in RoundDuration, when
rounding to years, months, or weeks, we need to divide by the length of
the year, month, or week in days. Throw an exception if this would
otherwise result in division by zero.

This is marked 'editorial' because division by zero with mathematical
values is an editorial error in the spec. However, the month/week
divisions didn't exist before the normative change in the previous commit,
so this would previously only have applied to years.

Closes: #2335

UPSTREAM_COMMIT=ed85e96910c2a1767828db344c4b5bf176560c15