Normative: Fix Duration rounding relative to ZonedDateTime
When the time portion of a duration, rounded relative to a ZonedDateTime,
would land on a non-24-hour day, we'd get incorrect results. This is a
regression from #2508 where although switching the order of
BalanceDateDurationRelative and BalanceTimeDurationRelative was correct,
we should not have removed the MoveRelativeZonedDateTime call.

Closes: #2742

UPSTREAM_COMMIT=b70e3494973120c586e1beceb213ac9803e8e545