Editorial: Use grammar parameterization to express whether 'Z' is allowed
In Plain types, the UTC designator 'Z' is not allowed in the string. This
was previously expressed with explicit checks in the ParseTemporal... AOs,
but can be more cleanly expressed with a [Z] parameter to the
DateTimeUTCOffset production.

UPSTREAM_COMMIT=c68f976ee301327a990dd409fa91be29b405e7f8