Editorial: Remove unnecessary CreateDataPropertyOrThrow
This is an unobservable object. From here it's passed to
DifferenceZonedDateTime, then DifferenceISODateTime, where it is copied
and the property largestUnit is overwritten. So nothing can ever actually
observe the value set here.

Thanks as usual to Anba for spotting it.

Closes: #2802

UPSTREAM_COMMIT=5cd78dc85590678b7a6282d17d01c697a2221282