Post-adjustment: Remove unnecessary object creation
This is to address belated feedback from Anba on #2718. The options object
previously created here is unnecessary, because 'constrain' is always the
default for this option. Not creating the options object here is
consistent with the rest of the spec text.

Closes: #2803

UPSTREAM_COMMIT=9e7a01ca0e4b68e392fb4ec86d9c9273a0300a4b