Editorial: Don't use ToIntegerOrInfinity for string parsing
ECMA-262 settled on ℝ(StringToNumber(CodePointsToString(foo))) as the
proper way to do this in ParseTimeZoneOffsetString. Adopt this in the
Temporal spec as well, with a similar note about precision.

This removes all uses of ToIntegerOrInfinity, so the issue about auditing
all uses of it to make sure they handle infinities correctly can be closed
as well.

Closes: #1566
Closes: #1467

(Rebase note: Some instances of ToIntegerOrInfinity were missed in a
previous rebase. This commit corrects that.)

UPSTREAM_COMMIT=4a2150b05dc81c24eefbfbb15ab9da389dd2e7b5