Polyfill: Import es-abstract deps directly in calendar.mjs
This is one step towards breaking the annoying circular import between
calendar.mjs and ecmascript.mjs.

(Rebase note: This is totally redundant in this polyfill, as we already
don't have es-abstract as a dependency. So this commit just changes the
way a few ecmascript.ts exports are imported in calendar.ts, to match
the upstream code.)

UPSTREAM_COMMIT=4867d9b89ffbc8e69ea81bb300431dfdb228f897