Editorial: Add AdjustDateDurationRecord AO
In a few places, we want to create a new Date Duration Record but with one
or more fields adjusted, as if Date Duration Records had a with() method.
This is easy to do with spreading and destructuring in JS, but is unwieldy
in spec language. Add an AdjustDateDurationRecord abstract operation to do
this more concisely.

See: #2308

UPSTREAM_COMMIT=f260a297fdae82a63a2ca00b4510702d5b1b33da