Polyfill: Calculate time zone UTC offset at nearest millisecond
We are getting our UTC offset information from Intl.DateTimeFormat which
doesn't distinguish between an epoch nanoseconds in the middle of a
millisecond or at a millisecond boundary. Drop the microsecond and
nanosecond when querying Intl.DateTimeFormat for the UTC offset.

This means, we assume the UTC offsets at these two times are always the
same:

  2024-09-25T16:47.123456789
  2024-09-25T16:47.123

We inline only the relevant parts of GetNamedTimeZoneDateTimeParts and
GetISOPartsFromEpoch into GetNamedTimeZoneOffsetNanoseconds, to avoid
converting back and forth between epoch nanoseconds and milliseconds and
avoid calling some extra legacy Date methods we don't need.

This allows eliminating the BalanceISODateTime / BalanceTime call, which
to my surprise was the slowest part of the slow test262 test
ZDT/p/getTimeZoneTransition/transition-at-instant-boundaries.js.

This knocks quite a chunk off the slow test262 test, because we avoid a
bunch of BigInt math.

UPSTREAM_COMMIT=ca3e6b67fa24cbd0043ec2d951654857abf90fe2