Editorial: Merge SubtractTimeDuration into AddTimeDuration
This previously needed to be a separate operation because you can't just
negate a record. Now that time durations are a number, you can negate
them, so no need for a subtraction operation when we already have an
addition operation.

UPSTREAM_COMMIT=36f1baa1dc5595b929b6969314b45a0309661ed5