Editorial: Avoid hitting assertion in DifferenceISODateTime
DifferenceISODateTime asserts that the ISO Date-Time Records passed to it
are within limits. Previously there were two places where this was not
guaranteed to be the case. Add checks that throw RangeError so the
assertion is not hit.

Closes: #3015

UPSTREAM_COMMIT=b5da334a17ede21e99da0d1d91a72a801639c5e4