rebase-tool: Allow 'continue' and 'finish' to work in exec step
If using the --exec argument, the rebase may stop during an exec step if,
for example, the tests failed. When stopped during an exec step, there's
no .git/rebase-merge/stopped-sha file, so the tool cannot pick the rebase
back up nor finish early.

Instead, get the SHA from .git/rebase-merge/done, if the last entry in the
file is an 'edit' step. If the last entry is instead an 'exec' step,
handle that separately.