Polyfill: More conformant implementation of Duration.p.toLocaleString
The previous implementation in the reference code was non-conformant on
hosts with Intl.DurationFormat implemented without Temporal.Duration
integration. On those hosts, the argument to format() would be treated as
a Duration-like property bag and its properties would be observably read,
even if it was a Temporal.Duration. The correct behaviour would be to get
the information from the Temporal.Duration object's internal slots.

UPSTREAM_COMMIT=cf955929dc04d8c83cc4cd77038c570ebb5d769a