Fix the ts-prune workarounds
I learned today that TypeScript treats everything in a .d.ts file as if it
was exported, whether it says `export` or not, unless you add an
`export {}` declaration somewhere in the file. See
https://github.com/microsoft/TypeScript/issues/57764

We had a bunch of ts-prune-ignore-next declarations in internaltypes.d.ts
to work around this. Adding `export {}` allows ts-prune to work as
expected without workarounds.