Draft: Use 128-bit floats instead of BigInts
I read a paper on double-double precision arithmetic (i.e., using two
Numbers to implement 128-bit floats with the same range as Number but
twice the precision.) This precision would allow safe integers covering
the entire epoch nanoseconds range that we support, as well as being
precise enough for all the floating point calculations that we do in
methods like Duration.p.total().

The paper is: Hida, Li, Bailey (2008). Library for double-double and
quad-double arithmetic. Technical report, Lawrence Berkeley National
Laboratory. https://www.davidhbailey.com/dhbpapers/qd.pdf

They published a C++ library at https://github.com/BL-highprecision/QD.
It was a quick experiment to implement a subset of the operations in JS
in lib/float128.ts. This allows us to get rid of both JSBI as well as
the power-of-10-math-by-string-manipulation in lib/math.ts.